backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
python-socketio==5.10.0
python-dotenv==1.0.0
pycryptodome==3.20.0
pyotp==2.9.0
qrcode==7.4.2
argon2-cffi==23.1.0
